conform with strict provenance
Previous version stored pointer as a `usize`, which causes warning in
Miri by default and errors with `-Zmiri-strict-provenance`.

This version stores pointers directly and allows using proper strict
provenance APIs with `sptr` feature. It potentially benefits
optimizations and clarifies semantics of pointer-to-integer casts.

This commit also rewrites a couple of `Clone` implementations to be
defined in terms of `Copy`.